+++ src/entities/reviews/index.ts
@@ -1,6 +1,7 @@
 export * from './model/types';
 export {getBestReviews} from './apis/api-service';
 export {default as ReviewList} from './ui/ReviewList';
-export {default as useSearchReviewsWithKeyword} from './model/useSearchReviewsWithQuery';
 export {default as ReviewArticle} from './ui/ReviewArticle';
 export {default as ReviewArticleLoading} from './ui/ReviewArticleLoading';
+export {default as useSearchReviews} from './model/useSearchReviews';
+export {default as useSearchReviewsWithKeyword} from './model/useSearchReviewsWithQuery';
