+++ src/features/review-filtering/index.ts
@@ -1,2 +1,3 @@
 export {default as CategoryBar} from './ui/CategoryBar';
 export {useSelectCategory} from './lib/useSelectCategory';
+export {useSelectCategoryFromUrl} from './lib/useSelectCategoryFromUrl';
