+++ src/features/review-sorting/index.ts
@@ -1,1 +1,4 @@
+export * from './model/type';
 export {default as SelectSortOptions} from './ui/SelectSortOptions';
+export {default as useSelectSortOption} from './lib/useSelectSortOption';
+export {default as isSortKey} from './lib/isSortKey';
